backend/requirements.txt
flask
flask_cors
requests
PyYAML
influxdb_client
paramiko